+++ CHANGELOG.md
@@ -4,6 +4,20 @@
 >LoginRadius Node JS SDK Change Log provides information regarding what has changed, more specifically what changes, improvements and bug fix has been made to the SDK. For more details please refer to the [LoginRadius API Documention](https://docs.loginradius.com/api/v2/sdk-libraries/nodejs)
 =============================================================
 
+Version 4.1.0
+
+Released on May 12, 2017
+<br>
+- Add start and end time for SOTT to validate it for long time frame.
+- Add API to get server time for SOTT if don't pass the start and end time.
+- Improved SOTT feature.
+- Add Email prompt auto login APIs.
+- Add Role Context and additional permission APIs
+
+See the documentation [here](https://docs.loginradius.com/api/v2/sdk-libraries/nodejs)
+
+=============================================================
+
 Version 4.0.0
 
 Released on April 14, 2017